Just playing around with the GUI interface....
Trying to give the GUI the option to choose stable vs. bleeding updates and then to have the action actually call the update.py script.....though this is not the preferred method.  Will need to write the update code from within this system.py script.  The action check_for_update will be where this code will be called from further down the the page.  This is not ready for integration by any means...however it will give you an idea of where to start.